Fix all "declaration after statement" warnings, for greater C89 compatibility.
Also, add -Werror in CI, so we don't forget about warnings in future.

Note that sjlj.c requires _FORTIFY_SOURCE to be neutered, or else the
fortifications interfere with the stack wizardry. Redefining _FORTIFY_SOURCE
causes a warning, and -Werror would convert that to an error, so we can't
use -Werror when testing sjlj.c.